v3.36.2: Fase 18.3 — Refactor + qualidade (A1 decompor executeCommand + A4 zero warnings + A5 Reserved.h map + A7 broken-pipe observability)
A1: AppManager_Commands.cpp 952L -> 624L (-35%). Os 8 cases mais longos
do switch/51 cases (>=30L cada: SENSOR_FIELD, ACCEPT_SENSOR, USER_ADD,
RESET_ADMIN, SET_TIME, IP_CFG, SET_DNS_CFG, USER_PASS) foram extraidos
para AppManager_CmdHandlers.cpp (novo, 336L). Cada case agora e uma
linha curta cmdHandleX(cmd, cfg, changed). Semantica identica.

A4: zero warnings em codigo proprio. Build full sem cache revelou 22
sites unicos (cache mascarava 13). Fixados: misleading-indentation
(Settings 110/280, Touch 726/1011/1012, Graph 736/742), unused vars
(startIdx, timeW, w, tBuf, pageColor, cfg), comment-within-comment
(Themes 272, Telemetry 1269), sign-compare (Auth 76, Touch 811),
missing-initializer (tm = {0} -> {}), format %d vs int32_t.
Bug real corrigido: field == "tmin" em CMD_SENSOR_FIELD comparava
ponteiros (char* vs literal addr) -- ramo NUNCA era alcancado, caia
em "Campo desconhecido". Agora strcmp().

A5: SystemDefs_Reserved.h (novo) com RESERVED_*_OFFSET constexpr para
os 7 overlays em SystemConfig::reserved[64] (Touch/Sound/Display/Cli/
Web/Setup/NetTime) + static_assert(RESERVED_FREE_OFFSET <= 64).
Coexiste com WEB_CONFIG_OFFSET/CLI_CONFIG_OFFSET legados.

A7: observability de broken pipe. maybeLogClientDisconnect() com
throttle 5s nos 4 overloads de safeSend cobre 30+ call sites sem
tocar handlers. WEB_CLIENT_DISCONNECT=575. Origin tag s/sN/sStr/sP +
early/post rastreia qual streaming morreu.

Validado em HW: 9/9 smoke tests pass (perms/status/config/calib OK,
heap 58KB, uptime valido, M2 traversal ainda rejeitado, device
estavel apos broken pipe induzido). Flash 98.2% (+0.1% boilerplate).

+++ DisplayManager_Auth.cpp
@@ -73,7 +73,7 @@ void DisplayManager::scrambleKeys() {
     const char poolUpper[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
     const char poolLower[] = "abcdefghijklmnopqrstuvwxyz";
     const char poolSpec[] = "!@#$%^&*()_+-=[]{}|;':\",./<>?\\~";
-    char expected = '\0'; if (_authStep < _expectedPin.length()) { expected = _expectedPin[_authStep]; }
+    char expected = '\0'; if ((size_t)_authStep < _expectedPin.length()) { expected = _expectedPin[_authStep]; }
     int expectedType = -1;
     if (expected >= '0' && expected <= '9') expectedType = 0;
     else if (expected >= 'A' && expected <= 'Z') expectedType = 1;
